requirements.txt
astroid==2.0.4
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
coveralls==1.1
defusedxml==0.5.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.6
django-cors-headers==1.3.1
django-extensions==2.1.0
django-filter==2.0.0
django-js-asset==1.1.0
django-mptt==0.9.1
django-rest-framework-docs==0.1.7
django-rest-swagger==2.1.0
django-toolbelt==0.0.1
djangorestframework==3.8.2
docopt==0.6.2
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
jsonpickle==1.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==4.3.0
oauthlib==2.1.0
openapi-codec==1.3.2
pluggy==0.7.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.5
py==1.5.4
pycodestyle==2.4.0
Pygments==2.2.0
PyJWT==1.6.4
pylint==2.1.0
python3-openid==3.1.0
pytz==2018.5
requests==2.20.0
requests-oauthlib==1.0.0
simplejson==3.16.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
static3==0.7.0
typed-ast==1.1.0
typing==3.6.4
uritemplate==3.0.0
urllib3==1.23
whitenoise==4.0
wrapt==1.10.11
